ENGR00116947: MX51: Fix the audio-pop/channel swapping issue in lp-apm mode.
In LPAPM mode, peripheral domain is clocked from the 24MHz clock.
There is a brief period of time when IPG clock drops close to 2.4MHz,
which violates the SSI to IPG clock ratio. And this causes incorrect data
to be clocked out of SSI.
Fix the code such that IPG clock does not go below 10MHz when
entering/exiting LPAPM mode.

Signed-off-by: Ranjani Vaidyanathan-RA5478 <Ranjani.Vaidyanathan@freescale.com>