ENGR00122851 WIFI: Faided to use WiFi SDIO card at SD slot0
The WP bit is set in the 4Bytes wifi write operations
accidentally.
That would cause the INT wouldn't be generated.

There is no such kind issue in SD/MMC write operations.
Since that the BUS load on wifi maybe much lighter than
that when SD/MMC is used.

Remove the 100K pull up pad configuration on the eSDHC1
WP pin can solve this issue.
Or
Remove the SION configuration can fix this issue too.
In the end, the second method is used.

Unify the platform data configurations, and enable the registers
dump when there are errors in the execution of esdhc driver.

Signed-off-by: Richard Zhu <r65037@freescale.com>