sched: Add a generic notifier when a task struct is about to be freed
This patch adds a notifier which can be used by subsystems that may
be interested in when a task has completely died and is about to
have it's last resource freed.

The Android lowmemory killer uses this to determine when a task
it has killed has finally given up its goods.

Change-Id: I4dfcdad3917734f86c16cdc75ee0db0d92f6d34f
Signed-off-by: San Mehat <san@google.com>